Implémentation de update delete ajouter un rating, recalculer la note average

+++ mvg-backend/controllers/books.js
@@ -1,6 +1,8 @@
 const Book = require('../models/book')
+const fs = require('fs');
 
-exports.createBook = (req, res) => {
+
+exports.createBook = (req, res, next) => {
   const bookObject = JSON.parse(req.body.book)
   delete bookObject.userId
 
@@ -10,11 +12,10 @@ exports.createBook = (req, res) => {
     imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   })
 
-  book.save()
-    .then(() => { res.status(201).json({message: 'Le livre a bien été créé'})})
-    .catch( error => res.status(400).json({error}))
 
-  console.log(book)
+  book.save()
+    .then(() => { res.status(201).json({book: book, message: 'Le livre a bien été créé'})})
+    .catch( error => res.status(400).json({error}))
 }
 
 exports.getAllBooks = (req, res) => {
@@ -50,27 +51,44 @@ exports.updateOneBook = (req, res) => {
       if (book.userId != req.auth.userId) {
         res.status(401).json({message: "Vous n'êtes pas autorisé à modifier le livre"})
       } else {
+        // Rajouté pour pas accumuler les photos updatées
+        const filename = book.imageUrl.split('/images/')[1];
+
+        fs.unlink(`images/${filename}`, () => {
         Book.updateOne({_id: req.params.id}, {...bookObject, _id: req.params.id})
           .then(() => res.status(200).json({ message: 'Le livre a bien été mis à jour' }))
           .catch(error => res.status(400).json('Error:' + error));
+        })
       }
-
     })
     .catch( error => res.status(400).json({error}))
 }
 
 exports.deleteOneBook = (req, res) => {
-  Book.deleteOne({_id: req.params.id})
-    .then(() => res.status(200).json({message: 'Le livre a bien été supprimé'}))
-    .catch(error => res.status(400).json('Error:' + error));
+  Book.findOne({ _id: req.params.id})
+    .then( book => {
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({message: "Vous n'êtes pas autorisé à supprimer le livre"})
+      } else {
+        const filename = book.imageUrl.split('/images/')[1];
+
+        fs.unlink(`images/${filename}`, () => {
+        Book.deleteOne({_id: req.params.id})
+          .then(() => res.status(200).json({message: 'Le livre a bien été supprimé'}))
+          .catch(error => res.status(400).json('Error:' + error));
+        })
+      }
+    })
+    .catch( error => res.status(400).json({error}))
 }
 
 exports.rateOneBook = async (req, res) => {
-  // Token
   try {
   // Dans le body je récupère le nouveau rating
-
-    const ratedBook = await Book.findOne({_id: req.params.id})
+  console.log(req.body)
+  console.log(req.params.id)
+  const ratedBook = await Book.findOne({_id: req.params.id})
+  console.log(ratedBook)
 
     if (!ratedBook) {
       return res.status(404).json({ error: 'Book not found' });
@@ -78,19 +96,36 @@ exports.rateOneBook = async (req, res) => {
 
     const ratedIds = ratedBook.ratings.map(rating => rating.userId)
 
-    if (ratedIds.includes(req.body.userId)) {
+    if (ratedIds.includes(req.auth.userId)) {
       return res.status(400).json({ error: 'You have already rated this book' });
 
     } else {
-      // à voir si l'array de rating est défini comme ça
-      ratedBook.ratings.push(req.body)
+      // Je construis un array en protégeant le userId qui note grâce au jwt
+      const rating = {
+        userId: req.auth.userId,
+        grade: req.body.rating
+      }
 
-      await ratedBook.save(); // Save the changes to the database
+      // Faut il utiliser la méthode updateOneBook? Je ne pense pas car c'est vraiment pour le user qui a posté le livre
+      ratedBook.ratings.push(rating)
 
-      return res.status(200).json({ message: 'Rating added successfully' });
+      // Lancer le calcul de averageRating à ce moment là ?
+      const allCurrentGrades = ratedBook.ratings.map(rating => rating.grade)
+      const newAverage = allCurrentGrades.reduce((acc, rating) => acc + rating) / allCurrentGrades.length
+      console.log(newAverage)
+
+      ratedBook.averageRating = newAverage
+
+      console.log(ratedBook)
+
+      ratedBook.save() // Save the changes to the database
+        .then(book => {
+
+          return res.status(200).json({book: book, message: 'Rating added successfully'})
+        })
+        .catch( error => res.status(400).json({error}))
     }
 
-    // Lancer le calcul de averageRating à ce moment là ?
 
   } catch (error) {
     console.error(error);
